[GOBBLIN-750] remove usage of depricated DatasetResolver and DatasetResolverFactory
