--- gobblin-api/src/main/java/org/apache/gobblin/dataset/DatasetResolver.java
@@ -1,54 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.gobblin.dataset;
-
-import org.apache.gobblin.configuration.State;
-
-
-/**
- * A {@link DatasetResolver} resolves job specific dataset
- *
- * @deprecated use the more general {@link DescriptorResolver}
- */
-@Deprecated
-public interface DatasetResolver extends DescriptorResolver {
-  /**
-   * Given raw Gobblin dataset, resolve job specific dataset
-   *
-   * @param raw a dataset in terms of Gobblin
-   * @param state configuration that helps resolve job specific dataset
-   * @return resolved dataset for the job
-   */
-  DatasetDescriptor resolve(DatasetDescriptor raw, State state);
-
-  @Override
-  default Descriptor resolve(Descriptor raw, State state) {
-    DatasetDescriptor rawDataset;
-
-    if (raw instanceof DatasetDescriptor) {
-      rawDataset = (DatasetDescriptor) raw;
-    } else if (raw instanceof PartitionDescriptor) {
-      rawDataset = ((PartitionDescriptor) raw).getDataset();
-    } else {
-      // type not supported
-      return null;
-    }
-
-    return resolve(rawDataset, state);
-  }
-}
