+++ gobblin-api/src/main/java/org/apache/gobblin/dataset/HiveToHdfsDatasetResolverFactory.java
@@ -20,9 +20,9 @@ package org.apache.gobblin.dataset;
 import com.typesafe.config.Config;
 
 
-public class HiveToHdfsDatasetResolverFactory implements DatasetResolverFactory {
+public class HiveToHdfsDatasetResolverFactory implements DescriptorResolverFactory {
   @Override
-  public DatasetResolver createResolver(Config config) {
+  public DescriptorResolver createResolver(Config config) {
     return HiveToHdfsDatasetResolver.INSTANCE;
   }
 }
